Oh, Workspace can contain multiple projects.
Add an existing project to a workspace to add a product to a multiproduct development endeavor.

https://developer.apple.com/library/ios/recipes/xcode_help-structure_navigator/articles/Adding_an_Existing_Project_to_a_Workspace.html

Adding an Existing Project to a Workspace

In the project navigator, Control-click in the empty space below the list.

Choose Add Files to “<Workspace_Name>”.

Navigate into the folder containing the .xcodeproj file to add.

Select the .xcodeproj file and click Add.